[REFACTOR] #3 ViewModel로 분리
